Adicionando layout de image comic junto com Fragment e button para voltar

+++ app/src/main/java/com/example/desafio03digitalhouse/home/view/main/HomeFragment.kt
@@ -50,7 +50,7 @@ class HomeFragment : Fragment() {
 
         myadapter = ComicAdapter(listaPersonagem){
 
-            val bundle = bundleOf("id" to it.id)
+            val bundle = bundleOf("id" to it.id.toInt())
 
             val navigation = Navigation.findNavController(view)
